Change to common dependencies

+++ build_scripts/tbs-os-module.sh
@@ -1,8 +1,22 @@
 #!/bin/bash
 
+##Set branch to pull from for dependencies
+set -ea
+
+: "${DEPENDENCY_BRANCH:=master}"
+
 ##Pull variables from github
-wget -nc https://raw.githubusercontent.com/CHBMB/Unraid-DVB/master/build_scripts/variables.sh
-. "$(dirname "$(readlink -f ${BASH_SOURCE[0]})")"/variables.sh
+wget -nc https://raw.githubusercontent.com/linuxserver/Unraid-Dependencies/${DEPENDENCY_BRANCH}/build_scripts/variables.sh
+wget -nc https://raw.githubusercontent.com/linuxserver/Unraid-Dependencies/${DEPENDENCY_BRANCH}/build_scripts/dvb-variables.sh
+
+source ./variables.sh
+
+if [[ -z "$D" ]]; then
+    echo "Must provide D in environment" 1>&2
+    exit 1
+fi
+
+source ${D}/dvb-variables.sh
 
 ##Restore /lib/modules/
 rm -rf  /lib/modules
